cloud_analyzer.common/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-analyzer-common"
version = "0.1.0"
description = "Multi-cloud cost optimization analyzer - Common Library"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Cloud Analyzer Team"}
]
dependencies = [
    "pydantic>=2.5.0",
    "boto3>=1.34.0",
    "azure-mgmt-compute>=30.0.0",
    "azure-mgmt-storage>=21.0.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-identity>=1.15.0",
    "google-cloud-compute>=1.14.0",
    "google-cloud-storage>=2.13.0",
    "google-cloud-resource-manager>=1.11.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.111.0",
    "msal>=1.26.0",
    "aiohttp>=3.9.1",
    "authlib>=1.3.0",
    "cryptography>=41.0.7",
    "python-dateutil>=2.8.2",
    "typing-extensions>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    "isort>=5.13.0",
    "pre-commit>=3.6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"